requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.5.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.5
gunicorn>=20.1.0
python-dotenv>=1.1.0
drf-yasg>=1.21.5
psutil>=5.9.0
python-socketio>=5.7.2
uvicorn>=0.20.0
websockets>=11.0.0
PyJWT>=2.6.0
asgiref>=3.6.0
sqlparse>=0.4.3
pytz>=2022.7
six>=1.16.0
markdown>=3.4.1
pygments>=2.15.0
python-engineio>=4.4.0
anyio>=3.6.2
Pillow>=10.0.0
h11>=0.14.0
idna>=3.4
sniffio>=1.3.0
starlette>=0.27.0
httptools>=0.5.0
pyyaml>=6.0
watchfiles>=0.19.0
channels>=4.0.0
channels-redis>=4.1.0
pytest>=8.3.5
pylint>=3.3.7
dotenv>=0.9.9
# Network scanning dependencies
scapy>=2.5.0
netaddr>=0.8.0
netifaces>=0.11.0
python-nmap>=0.7.1
pyroute2>=0.7.3
django-filter>=25.1